+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [1.1.4] - 2021-08-20
+### Changed
+- Replaced custom `HashTable` with `ManyKeysMap`
+
 ### Added
 - `Pivot` and `PivotTable` have a new `reducerInitialValue` prop to allow complex reducers (same API as JS `Array.reduce()`)